fix(genesis): clean up chain config
