+++ crates/genesis/src/chain/superchain.rs
@@ -1,0 +1,18 @@
+//! Contains the superchain level.
+
+/// Level of integration with the superchain.
+#[derive(Debug, Copy, Clone, Default, Hash, Eq, PartialEq)]
+#[cfg_attr(any(test, feature = "arbitrary"), derive(arbitrary::Arbitrary))]
+#[cfg_attr(feature = "serde", derive(serde_repr::Serialize_repr, serde_repr::Deserialize_repr))]
+#[repr(u8)]
+pub enum SuperchainLevel {
+    /// Frontier chains are chains with customizations beyond the
+    /// standard OP Stack configuration and are considered "advanced".
+    Frontier = 0,
+    /// A candidate for a standard chain.
+    #[default]
+    StandardCandidate = 1,
+    /// Standard chains don't have any customizations beyond the
+    /// standard OP Stack configuration and are considered "vanilla".
+    Standard = 2,
+}
